# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FSP Talent API",
  "description": "\n**FSP Talent** — платформа подбора ИТ-специалистов с обратной механикой:\nработодатель сам находит кандидата в категоризированной базе и отправляет приглашение с вилкой ЗП.\n\n### Как авторизоваться в Swagger\n1. `POST /api/auth/register` → `POST /api/auth/verify-email` (в dev-режиме токен приходит в ответе регистрации).\n2. Кнопка **Authorize** вверху: username = email, password = пароль.\n\n### Формат ошибок\n`{\"code\": \"MACHINE_CODE\", \"message\": \"текст для пользователя\", \"detail\": \"...\"}`, для 422 ещё `fields: {поле: ошибка}`\n",
  "version": "1.3.0"
 },
 "paths": {
  "/api/auth/register": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Регистрация кандидата или работодателя",
    "description": "Создаёт пользователя и пустой профиль нужной роли, отправляет письмо для подтверждения email.",
    "operationId": "register_api_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RegisterResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "429": {
      "description": "Слишком много запросов",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/verify-email": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Подтвердить email (токен из письма)",
    "operationId": "verify_email_api_auth_verify_email_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VerifyEmailRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/resend-verification": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Отправить письмо подтверждения ещё раз",
    "operationId": "resend_api_auth_resend_verification_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResendVerificationRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Message"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/login": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Вход (JSON) — для фронтенда",
    "description": "После 5 неудачных попыток подряд вход по этому email блокируется на 15 минут (429).",
    "operationId": "login_api_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenPair"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "429": {
      "description": "Слишком много запросов",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/token": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Вход через форму — для кнопки Authorize в Swagger (username = email)",
    "operationId": "login_form_api_auth_token_post",
    "requestBody": {
     "content": {
      "application/x-www-form-urlencoded": {
       "schema": {
        "$ref": "#/components/schemas/Body_login_form_api_auth_token_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenPair"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "429": {
      "description": "Слишком много запросов",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/refresh": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Обновить пару токенов (ротация)",
    "description": "Старый refresh-токен «сгорает», выдаётся новая пара — сохраните ОБА новых токена.\nПовторное использование сгоревшего токена = признак кражи -> вся сессия отзывается (401 TOKEN_REUSED).",
    "operationId": "refresh_api_auth_refresh_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RefreshRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenPair"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/logout": {
   "post": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Выход (отзыв сессии)",
    "description": "С `refresh_token` — завершает эту сессию; без тела — выход на всех устройствах.\nПосле выхода refresh-токен больше не обменять. Фронт удаляет оба токена у себя.",
    "operationId": "logout_api_auth_logout_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/LogoutRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Data"
       }
      }
     }
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/auth/me": {
   "get": {
    "tags": [
     "Авторизация"
    ],
    "summary": "Кто я + чек-лист онбординга",
    "operationId": "me_api_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_login_form_api_auth_token_post": {
    "properties": {
     "grant_type": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^password$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grant Type"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "format": "password",
      "title": "Password"
     },
     "scope": {
      "type": "string",
      "title": "Scope",
      "default": ""
     },
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id"
     },
     "client_secret": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "format": "password",
      "title": "Client Secret"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "Body_login_form_api_auth_token_post"
   },
...
   "ErrorResponse": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "items": {},
        "type": "array"
       }
      ],
      "title": "Detail"
     },
     "fields": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "string"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fields"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "detail"
    ],
    "title": "ErrorResponse"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest"
   },
   "LogoutRequest": {
    "properties": {
     "refresh_token": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Refresh Token",
      "description": "Завершить эту сессию. Без токена — выйти на всех устройствах"
     }
    },
    "type": "object",
    "title": "LogoutRequest"
   },
   "MeOut": {
    "properties": {
     "user": {
      "$ref": "#/components/schemas/UserOut"
     },
     "candidate_profile_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidate Profile Id"
     },
     "employer_profile_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employer Profile Id"
     },
     "onboarding": {
      "additionalProperties": true,
      "type": "object",
      "title": "Onboarding",
      "description": "Что ещё не сделано: для подсказок в UI"
     }
    },
    "type": "object",
    "required": [
     "user"
    ],
    "title": "MeOut"
   },
   "Message": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "message"
    ],
    "title": "Message"
   },
...
   "RefreshRequest": {
    "properties": {
     "refresh_token": {
      "type": "string",
      "title": "Refresh Token"
     }
    },
    "type": "object",
    "required": [
     "refresh_token"
    ],
    "title": "RefreshRequest"
   },
   "RegisterRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email",
      "examples": [
       "ivan@example.ru"
      ]
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 8,
      "title": "Password",
      "examples": [
       "StrongPass123"
      ]
     },
     "role": {
      "type": "string",
      "enum": [
       "candidate",
       "employer"
      ],
      "title": "Role",
      "description": "Какой кабинет создать"
     },
     "full_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name",
      "description": "Для кандидата (можно позже)"
     },
     "company_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Company Name",
      "description": "Для работодателя: можно сразу или позже в профиле компании"
     },
     "consent_processing": {
      "type": "boolean",
      "title": "Consent Processing",
      "description": "Галочка «Даю согласие на обработку ПДн» на форме регистрации",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "email",
     "password",
     "role"
    ],
    "title": "RegisterRequest"
   },
   "RegisterResponse": {
    "properties": {
     "user_id": {
      "type": "integer",
      "title": "User Id"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "verification_required": {
      "type": "boolean",
      "title": "Verification Required"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "dev_verification_token": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dev Verification Token",
      "description": "Только в режиме dev: токен подтверждения, чтобы не ходить в почту на демо"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "email",
     "role",
     "verification_required",
     "message"
    ],
    "title": "RegisterResponse"
   },
   "ResendVerificationRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     }
    },
    "type": "object",
    "required": [
     "email"
    ],
    "title": "ResendVerificationRequest"
   },
...
   "TokenPair": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "refresh_token": {
      "type": "string",
      "title": "Refresh Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In",
      "description": "Через сколько секунд истечёт access_token"
     },
     "role": {
      "type": "string",
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "refresh_token",
     "expires_in",
     "role"
    ],
    "title": "TokenPair"
   },
   "UserOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "is_email_verified": {
      "type": "boolean",
      "title": "Is Email Verified"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "role",
     "is_email_verified",
     "created_at"
    ],
    "title": "UserOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VerifyEmailRequest": {
    "properties": {
     "token": {
      "type": "string",
      "title": "Token"
     }
    },
    "type": "object",
    "required": [
     "token"
    ],
    "title": "VerifyEmailRequest"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/token"
     }
    }
   }
  }
 }
}